+++ Sources/Gf/matrix3d.cpp
@@ -452,7 +452,7 @@ GfVec3d GfMatrix3d::DecomposeRotation(const GfVec3d &axis0,
 bool GfIsClose(GfMatrix3d const &m1, GfMatrix3d const &m2, double tolerance) {
   for (size_t row = 0; row < 3; ++row) {
     for (size_t col = 0; col < 3; ++col) {
-      if (!GfIsClose(m1[row][col], m2[row][col], tolerance))
+      if (!GfIsClose(m1[static_cast<int>(row)][col], m2[static_cast<int>(row)][col], tolerance))
         return false;
     }
   }
